pyproject.toml
[project]
name = "distributed-kv-cache"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "protobuf>=6.33.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
